feat: Add Cloud Recording client implementation and update project structure
- Introduced CloudRecordingClient and its implementation (CloudRecordingClientImpl) for managing cloud recording functionalities.
- Added CloudRecordingConfig class for configuration management.
- Implemented individual and mix recording scenarios with respective request and response classes.
- Updated pom.xml to include the Maven resources plugin for copying the LICENSE file.
- Added LICENSE file for project compliance.
- Refactored existing classes to improve structure and readability, including enhanced documentation for methods and parameters.
