+++ agora-rest-client-core/src/main/java/io/agora/rest/services/cloudrecording/api/req/UpdateResourceReq.java
@@ -144,11 +144,20 @@ public class UpdateResourceReq {
         }
     }
 
+    /**
+     * @brief Update subscription lists.
+     * @since v0.4.0
+     */
     public static class StreamSubscribe {
-
+        /*
+         * The audio subscription list.(Optional)
+         */
         @JsonProperty("audioUidList")
         private AudioUIDList audioUidList;
 
+        /**
+         * The video subscription list.(Optional)
+         */
         @JsonProperty("videoUidList")
         private VideoUIDList videoUidList;
 
@@ -210,11 +219,30 @@ public class UpdateResourceReq {
         }
     }
 
+    /**
+     * @brief Update audio subscription list.
+     * @since v0.4.0
+     */
     public static class AudioUIDList {
 
+        /**
+         * Specify which UIDs' audio streams to subscribe to.
+         * 
+         * If you want to subscribe to the audio stream of all UIDs, no need to set
+         * this field.
+         * 
+         * Set as ["#allstream#"] to subscribe to the audio streams of all UIDs in
+         * the channel.
+         */
         @JsonProperty("subscribeAudioUids")
         private List<String> subscribeAudioUIDs;
 
+        /**
+         * Specify which UIDs' audio streams not to subscribe to.
+         * 
+         * The cloud recording service will subscribe to the audio streams of all other
+         * UIDs except the specified ones.
+         */
         @JsonProperty("unSubscribeAudioUids")
         private List<String> unsubscribeAudioUIDs;
 
@@ -276,11 +304,30 @@ public class UpdateResourceReq {
         }
     }
 
+    /**
+     * @brief Update video subscription list.
+     * @since v0.4.0
+     */
     public static class VideoUIDList {
 
+        /**
+         * Specify which UIDs' video streams to subscribe to.
+         * 
+         * If you want to subscribe to the video stream of all UIDs, no need to set this
+         * field.
+         * 
+         * Set as ["#allstream#"] to subscribe to the video streams of all UIDs in the
+         * channel.
+         */
         @JsonProperty("subscribeVideoUids")
         private List<String> subscribeVideoUIDs;
 
+        /*
+         * Specify which UIDs' video streams not to subscribe to.
+         * 
+         * The cloud recording service will subscribe to the video streams of all other
+         * UIDs except the specified ones.
+         */
         @JsonProperty("unSubscribeVideoUids")
         private List<String> unsubscribeVideoUIDs;
