+++ agora-rest-client-core/src/main/java/io/agora/rest/services/cloudrecording/api/res/UpdateLayoutResourceRes.java
@@ -2,16 +2,33 @@ package io.agora.rest.services.cloudrecording.api.res;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+/**
+ * @brief Returned by the mix recording UpdateLayout API.
+ * @since v0.4.0
+ */
 public class UpdateLayoutResourceRes {
+    /**
+     * The name of the channel to be recorded.
+     */
     @JsonProperty("cname")
     private String cname;
 
+    /**
+     * The user ID used by the cloud recording service in the RTC channel to
+     * identify the recording service in the channel.
+     */
     @JsonProperty("uid")
     private String uid;
 
+    /**
+     * Unique identifier of the resource.
+     */
     @JsonProperty("resourceId")
     private String resourceId;
 
+    /**
+     * Unique identifier of the recording session.
+     */
     @JsonProperty("sid")
     private String sid;
