+++ agora-rest-client-core/src/main/java/io/agora/rest/services/cloudrecording/scenario/mix/res/QueryMixHLSAndMP4RecordingResourceRes.java
@@ -2,16 +2,36 @@ package io.agora.rest.services.cloudrecording.scenario.mix.res;
 
 import io.agora.rest.services.cloudrecording.api.res.QueryResourceRes;
 
+/**
+ * @brief Response returned by the mix recording QueryHLSAndMP4 API.
+ * @since v0.4.0
+ */
 public class QueryMixHLSAndMP4RecordingResourceRes {
-
+    /**
+     * The name of the channel to be recorded.
+     */
     private String cname;
 
+    /**
+     * The user ID used by the cloud recording service in the RTC channel to
+     * identify the recording service in the channel.
+     */
     private String uid;
 
+    /**
+     * Unique identifier of the resource.
+     */
     private String resourceId;
 
+    /**
+     * Unique identifier of the recording session.
+     */
     private String sid;
 
+    /**
+     * Server response. See
+     * {@link QueryResourceRes.MixRecordingHLSAndMP4ServerResponse}.
+     */
     private QueryResourceRes.MixRecordingHLSAndMP4ServerResponse serverResponse;
 
     public static Builder builder() {
