+++ agora-rest-client-core/src/main/java/io/agora/rest/services/cloudrecording/scenario/web/req/UpdateWebRecordingResourceClientReq.java
@@ -2,10 +2,21 @@ package io.agora.rest.services.cloudrecording.scenario.web.req;
 
 import io.agora.rest.services.cloudrecording.api.req.UpdateResourceReq;
 
+/**
+ * @brief Client request for updating web page recording.
+ * @since v0.4.0
+ */
 public class UpdateWebRecordingResourceClientReq {
 
+    /**
+     * Used to update the web page recording configurations.(Optional)
+     */
     private UpdateResourceReq.WebRecordingConfig webRecordingConfig;
 
+    /**
+     * Used to update the configurations for pushing web page recording to the
+     * CDN.(Optional)
+     */
     private UpdateResourceReq.RtmpPublishConfig rtmpPublishConfig;
 
     public static Builder builder() {
